Integrate new CSS for project
These changes implement a global form for the css containers to utilize throughout our project.

This was heavily influenced from tutorials online. We essentiall have a card-body that is centered in the screen. This essentially is the body of the form.

Then we have a form box which will essentially contain any input from the user.

Link to tutorial: https://jsfiddle.net/StartBootstrap/amxr8n19/