Allow user to create an empty Cardgame and to to lobby page feature This first seeds the database with the current_game attribute to 0 signaling it is NOT in a game.
In the feature I first select new game which is already in place. Then I create a game with N sinks, N decks, jokers, shuffled and set of cards in hand. For now this has no functionality but will soon. Finally it verifies we are on the lobby page after creating the game.