Notify other users when a user has left or started the game.
Originally, the user would have to restart the page if the game was to be started or ended.

Now when a single user starts a game, the game will be marked as started and Pusher will reload the page. When the page notices its started it will redirect accordingly.

The same goes for the end game. When a user ends the game it will reload all the pages for the users. It will then notice the page has ended and will end accordingly.