Display messages on game board
This will create a table html element for the messages. If the message came from the current user the username will be green. If the message came from another user, the username is in blue. The table elements will be listed and have the further most table on the bottom.

Co-authored-by: Bailey Leuth <bailey-leuth@uiowa.edu>